# api/openapi.yaml
openapi: 3.0.3
info:
  title: Official Account Service API
  version: 1.0.0
  description: WeChat official-account authorization, article, material, publishing, callback, dashboard, and task-queue APIs.
servers:
  - url: http://localhost:8080
...
paths:
...
  /wechat/component/callback:
    post:
      tags: [WeChatCallbacks]
      summary: Receive WeChat component callbacks
      parameters: [{$ref: "#/components/parameters/EncryptType"}, {$ref: "#/components/parameters/MsgSignature"}, {$ref: "#/components/parameters/Timestamp"}, {$ref: "#/components/parameters/Nonce"}]
      requestBody: {$ref: "#/components/requestBodies/XMLCallback"}
      responses: {"200": {$ref: "#/components/responses/CallbackOK"}, "400": {$ref: "#/components/responses/InvalidRequest"}}
...
  /api/v1/wechat/authorization-url:
    get:
      tags: [Authorization]
      summary: Generate a WeChat component authorization URL
      parameters:
        - {name: component_appid, in: query, required: true, schema: {type: string}}
        - {name: redirect_uri, in: query, required: true, schema: {type: string, format: uri}}
        - {name: auth_type, in: query, schema: {type: integer, enum: [1, 2, 3]}}
        - {name: biz_appid, in: query, schema: {type: string}}
      responses: {"200": {description: Authorization URL, content: {application/json: {schema: {$ref: "#/components/schemas/AuthorizationURL"}}}}, "400": {$ref: "#/components/responses/InvalidRequest"}, "501": {$ref: "#/components/responses/NotImplemented"}}
  /api/v1/wechat/authorization-callback:
    get:
      tags: [Authorization]
      summary: Handle WeChat authorization redirect
      parameters:
        - {name: tenant_id, in: query, required: true, schema: {type: string}}
        - {name: component_appid, in: query, required: true, schema: {type: string}}
        - {name: auth_code, in: query, required: true, schema: {type: string}}
      responses: {"200": {description: Authorized account without token fields, content: {application/json: {schema: {$ref: "#/components/schemas/Account"}}}}, "400": {$ref: "#/components/responses/InvalidRequest"}, "501": {$ref: "#/components/responses/NotImplemented"}}
...
  /api/v1/materials/inline-images:
    post:
      tags: [Materials]
      summary: Upload an article body image
      parameters: [{$ref: "#/components/parameters/TenantID"}]
      requestBody: {$ref: "#/components/requestBodies/MaterialUpload"}
      responses: {"201": {description: Uploaded material, content: {application/json: {schema: {$ref: "#/components/schemas/MaterialAsset"}}}}, "400": {$ref: "#/components/responses/InvalidRequest"}}
  /api/v1/materials/covers:
    post:
      tags: [Materials]
      summary: Upload an article cover image
      parameters: [{$ref: "#/components/parameters/TenantID"}]
      requestBody: {$ref: "#/components/requestBodies/MaterialUpload"}
      responses: {"201": {description: Uploaded material, content: {application/json: {schema: {$ref: "#/components/schemas/MaterialAsset"}}}}, "400": {$ref: "#/components/responses/InvalidRequest"}}
  /api/v1/articles/{id}/publish:
    post:
      tags: [Publishing]
      summary: Create a WeChat draft and submit it for publishing
      parameters: [{$ref: "#/components/parameters/TenantID"}, {$ref: "#/components/parameters/ID"}]
      responses: {"201": {description: Publish record, content: {application/json: {schema: {$ref: "#/components/schemas/PublishRecord"}}}}, "501": {$ref: "#/components/responses/NotImplemented"}}
  /api/v1/publish-records:
    get:
      tags: [Publishing]
      summary: List tenant publish records
      parameters: [{$ref: "#/components/parameters/TenantID"}]
      responses: {"200": {$ref: "#/components/responses/PublishRecordList"}}
    post:
      tags: [Publishing]
      summary: Create a local publish record
      parameters: [{$ref: "#/components/parameters/TenantID"}]
      requestBody: {$ref: "#/components/requestBodies/CreatePublishRecord"}
      responses: {"201": {description: Created publish record, content: {application/json: {schema: {$ref: "#/components/schemas/PublishRecord"}}}}, "400": {$ref: "#/components/responses/InvalidRequest"}}
  /api/v1/publish-records/{id}:
    get:
      tags: [Publishing]
      summary: Get a publish record
      parameters: [{$ref: "#/components/parameters/TenantID"}, {$ref: "#/components/parameters/ID"}]
      responses: {"200": {description: Publish record, content: {application/json: {schema: {$ref: "#/components/schemas/PublishRecord"}}}}, "404": {$ref: "#/components/responses/NotFound"}}
  /api/v1/articles/{id}/publish-records:
    get:
      tags: [Publishing]
      summary: List publish records for an article
      parameters: [{$ref: "#/components/parameters/TenantID"}, {$ref: "#/components/parameters/ID"}]
      responses: {"200": {$ref: "#/components/responses/PublishRecordList"}}
  /api/v1/publish-records/{id}/status:
    put:
      tags: [Publishing]
      summary: Update a local publish status
      parameters: [{$ref: "#/components/parameters/TenantID"}, {$ref: "#/components/parameters/ID"}]
      requestBody: {$ref: "#/components/requestBodies/UpdatePublishStatus"}
      responses: {"200": {description: Updated publish record, content: {application/json: {schema: {$ref: "#/components/schemas/PublishRecord"}}}}, "400": {$ref: "#/components/responses/InvalidRequest"}}
  /api/v1/publish-records/{id}/sync-status:
    post:
      tags: [Publishing]
      summary: Poll WeChat publish status and update local state
      parameters: [{$ref: "#/components/parameters/TenantID"}, {$ref: "#/components/parameters/ID"}]
      responses: {"200": {description: Synced publish record, content: {application/json: {schema: {$ref: "#/components/schemas/PublishRecord"}}}}, "501": {$ref: "#/components/responses/NotImplemented"}}
...
components:
  parameters:
    TenantID: {name: X-Tenant-ID, in: header, required: true, schema: {type: string}}
    ID: {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
...
    EncryptType: {name: encrypt_type, in: query, schema: {type: string, enum: [aes]}}
    MsgSignature: {name: msg_signature, in: query, schema: {type: string}}
    Timestamp: {name: timestamp, in: query, schema: {type: string}}
    Nonce: {name: nonce, in: query, schema: {type: string}}
  requestBodies:
    XMLCallback: {required: true, content: {application/xml: {schema: {type: string}}}}
...
    CreatePublishRecord: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreatePublishRecordRequest"}}}}
    UpdatePublishStatus: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/UpdatePublishStatusRequest"}}}}
    MaterialUpload:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            required: [authorizer_id, article_id, file]
            properties:
              authorizer_id: {type: integer, format: int64, minimum: 1}
              article_id: {type: integer, format: int64, minimum: 1}
              file: {type: string, format: binary}
  responses:
...
    CallbackOK: {description: Callback accepted, content: {text/plain: {schema: {type: string, example: success}}}}
    InvalidRequest: {description: Invalid request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    NotImplemented: {description: Dependency not configured, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
...
    PublishRecordList: {description: Publish record list, content: {application/json: {schema: {type: object, required: [items], properties: {items: {type: array, items: {$ref: "#/components/schemas/PublishRecord"}}}}}}}
...
  schemas:
...
    Error: {type: object, required: [error], properties: {error: {type: string, enum: [invalid_request, not_found, not_implemented, internal_error]}}}
    AuthorizationURL: {type: object, required: [authorization_url, pre_auth_code_expires_in_sec], properties: {authorization_url: {type: string, format: uri}, pre_auth_code_expires_in_sec: {type: integer}}}
    Account:
      type: object
      description: Safe account representation; token, secret, and refresh fields are intentionally omitted.
      properties: {id: {type: integer, format: int64}, tenant_id: {type: string}, app_id: {type: string}, name: {type: string}, avatar_url: {type: string}, status: {type: string, enum: [active, revoked, refresh_failed]}, last_synced_at: {type: string, format: date-time}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}
...
    MaterialAsset:
      type: object
      properties: {id: {type: integer, format: int64}, tenant_id: {type: string}, authorizer_id: {type: integer, format: int64}, article_id: {type: integer, format: int64}, usage: {type: string, enum: [inline_image, cover]}, local_url: {type: string}, wechat_url: {type: string}, media_id: {type: string}, created_at: {type: string, format: date-time}}
    PublishRecord:
      type: object
      properties: {id: {type: integer, format: int64}, tenant_id: {type: string}, authorizer_id: {type: integer, format: int64}, article_id: {type: integer, format: int64}, wechat_publish_id: {type: string}, wechat_article_id: {type: string}, status: {type: string, enum: [publishing, published, failed]}, error_code: {type: string}, error_message: {type: string}, submitted_at: {type: string, format: date-time}, finished_at: {type: string, format: date-time}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}
    CreatePublishRecordRequest: {type: object, required: [article_id], properties: {article_id: {type: integer, format: int64, minimum: 1}, wechat_publish_id: {type: string}}}
    UpdatePublishStatusRequest: {type: object, required: [status], properties: {status: {type: string, enum: [publishing, published, failed]}, wechat_article_id: {type: string}, error_code: {type: string}, error_message: {type: string}}}
...